requirements.txt
tensorflow==1.13.2
keras==2.1.5
sklearn
opencv-python
scipy
git+https://github.com/JiahuiYu/neuralgym
pillow
matplotlib
pyecharts